refactored test utils
