+++ tests/fixtures/databases/mealie/fixture_mealie_database.py
@@ -18,7 +18,7 @@ from AppLambda.src.models.mealie import (
     MealieShoppingListRecipeRef,
     Unit,
 )
-from tests.utils import random_bool, random_int, random_string
+from tests.utils.generators import random_bool, random_int, random_string
 
 from .mock_mealie_database import MockMealieDBKey, MockMealieServer
