+++ tests/utils/users.py
@@ -1,5 +1,3 @@
-import random
-import string
 from datetime import timedelta
 from typing import Optional
 
@@ -10,31 +8,7 @@ from AppLambda.src.routes import core
 from AppLambda.src.services.auth_token import AuthTokenService
 from AppLambda.src.services.user import UserService
 
-
-def random_string(length=10) -> str:
-    return "".join(random.choice(string.ascii_lowercase + string.digits) for _ in range(length)).strip()
-
-
-def random_email(length=10) -> str:
-    return "".join(random.choice(string.ascii_lowercase + string.digits) for _ in range(length)) + "@example.com"
-
-
-def random_password(length=20) -> str:
-    return random_string(length)
-
-
-def random_bool() -> bool:
-    return bool(random.getrandbits(1))
-
-
-def random_int(min=-4294967296, max=4294967296) -> int:
-    return random.randint(min, max)
-
-
-def random_url(https=True) -> str:
-    """all random URLs are the same length, with or without https (25 characters)"""
-
-    return f"{'https' if https else 'http'}://{random_string(5 if https else 6)}.example.com"
+from .generators import random_email, random_password
 
 
 def get_auth_headers(
